requirements.txt
mysql-connector
mysql-connector-python